requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
click==8.1.8
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
multidict==6.1.0
nltk==3.9.1
numpy==2.2.1
pandas==2.2.3
propcache==0.2.1
python-dotenv==1.0.1
regex==2024.11.6
scikit-learn==1.6.0
scipy==1.15.0
threadpoolctl==3.5.0
tqdm==4.67.1
yarl==1.18.3
